Create some basic test-cases for the plug-in and planners.
Here, we work through, document & get familiar with the relevant Spark's extension interfaces (because Spark is very big, we need to focus on very specific parts of it to make headway) and create dummy implementations of plan rewriters.

The first dummy implementation extracts a simple execution plan and the corresponding unit test pattern-matches against this execution plan to document the key important bits for reference from this unit test file.

The second dummy implementation rewrites the execution plan fully to a stubbed one that returns a constant result.

For the stubbed one, we create 2 unit tests:
- One that is passing because of the constant result matching the sum of the inputs
- One that is failing because of the constant result not matching the sum of the inputs

This TDD approach will lead us to develop the right plan rewriter, which will then interface with the Vector Engine. The strategic approach here is to do the necessary bottom up bits from the Spark side, and then the necessary bottom up bits from the Vector Engine side, and do the minimum necessary work to make some basic tests pass.

As part of the plan generation we will generate some sort of mini-program that can be sent to the VE code to do all the computation right there and then. However, details of this are yet to be worked out.