Evaluate the sum at the proper stage, as part of the RDD, rather than stage transformation.
This gives rise to parallelization powers.

- We remove now-redundant test from `SqlPluginTest`
- Create `SummingSparkPlan`, which sums the child's input digits. Note that this does not support things like push-down expressions yet; `SUM(x+2)` will return the same as `SUM(x)`. This is for later, as we may even want to do this part of the push-down into Aurora.
- `SummingSparkPlan.summingRdd` takes an input `RDD`, coalesces everything into 1 partition, and then performs the computation as part of the `RDD` `mapPartitions` method, to return 1 row in the end.
- `SumPlanExtractor` split into two parts, both of which we consume separately.
- Customise `SummingPlugin` so that it can use different summers where we need it to, rather than just the Nec summer.
- `SummingSparkPlanSpec` is a simple unit test for the summingRdd for `SummingSparkPlan`

Here is the Physical plan produced by the in-memory table:
```
== Physical Plan ==
SummingSparkPlan com.nec.spark.agile.BigDecimalSummer$$$Lambda$22692/657353307@28150ea1
+- LocalTableScan [value#1]
```

Here is the Physical plan produced by the CSV loader:

```
SummingSparkPlan com.nec.spark.agile.BigDecimalSummer$ScalaSummer$@44f12c96
+- FileScan csv [value#52] Batched: false, DataFilters: [], Format: CSV, Location: InMemoryFileIndex[file:/aurora4spark-parent/aurora4/.../.sample.csv, PartitionFilters: [], PushedFilters: [], ReadSchema: struct<value:decimal(38,18)
```

This demonstrates that we can rewrite and consume from another stage; so we can gulp all the different stages, and basically create a powerful compiled stage that we can simply send off to Aurora to compile, and pass the input in whatever way that is appropriate. We will be generating C from here, to be for example, passed into Python scripts like here https://github.com/SX-Aurora/py-veo/blob/master/examples/average.py
It is to be figured out later what protocol we will use to test over SSH, but it is just an implementation detail; we could even use sockets to communicate with a Python server over SSH, for example.
We can also simply just pass generated Python scripts straight over SSH if we need to.

The reason for using these SSH commands is for super fast development cycles (as opposed to creating a JAR, uploading it to the server, and running it there). Lightweight protocols ftw.