Integrate Dominik's runner script; Introduce a Bundle & BundleExecutor
Now we use a real C-based kernel in Dominik's script, which means we are one step closer to running the kernel from Scala. This time we do it via a Python script.

Goal is to make the acceptance tests pass. A few of them are failing; namely because we are returning a static number.

Classes:
- `Bundle`: the thing that we package and execute. There, we have the method that generates the self-contained bash script to run
- `BundleExecutor`: contains the classes needed to perform the execution of the bundle over SSH.
- `BundleExecutorSpec`: so far, we acceptance-test that we can run anything on Python remotely.