Update READMEs & set up SBT in the sub-directory.
Maven vs SBT for me is a difference of around 2-3x development speed because of TDD and compile times.

Isolating it to the very specific subproject; we will still continue to use Maven for the outer layer for ease of access.

READMEs: better to segregate to avoid noise from different places.

Also saw a test failure from Py script, fixed it