Support AVG()
This one was interesting: if you use decimal types for avg(), then the returned schema will have different scale than sum(), which caused lots of funky issues. However, it is easier to use Double type instead. I encoded the weird behaviour in unit tests so it is clear what is happening, rather than getting funny surprises.

Ensure that sum() does not conflict with avg() in terms of pattern matching.

Ensure the csv read works on Windows due to path changes.

Create `SparkSqlPlanExtension`, a more generic way to do an extension for our use cases, where we update the rules outside the extension. We are not abstracting this right now with the SumPlugin because we are not certain what the future holds in terms of the shape (it is less painful to copy & refactor later, than it is to try to fit something into a shape that is not appropriate).

Implement the avg() kernel.

Implement a pattern matcher for Avg()

Implemented also a `StaticNumberPlan` for testing and identifying any output issues.