Include Frovedis JARs as an Ivy repository
These should aid us in developing our plug-in.

It is natively available through SBT so that we could easily browse through all the test sources, and if need be, actually call into Frodevis.

Also included a Dockerfile to setup a Frodevis locally. What is cool here is that there is an x86 mode, which we could leverage for faster local development.