Hook the Aurora4j API into the Spark plug-in.
Updated `/README.md` for usage instructions.

Here, we compile with `ncc` once at the level of the driver.
Then, for each partition we call into the vector engine using the pre-compiled kernel. To that kernel we pass batches of data, compute the average, and then coalesce these batches into one final result.

We need to explicitly specify that we want 'offheap' to be used for columnar data - however we'll be able to use other approaches later.