Use an explicit `AddPlanExtractor`, include it in `LocalVeoExtension` and fix the naming.
We extract the creation of the VE process into the `Aurora4SparkExecutorPlugin`. Rename `SqlPlugin` to `AuroraSqlPlugin`.

Upon further research, the most obvious approach to do is to rewrite the Spark physical plan into something that we can comprehend on the VE side - however on the VE side the closest that we come to is Frovedis, however I foresee a good number of challenges in making this happen quickly. On top of it, this really means to rewrite *everything* and then derive something.

An alternative approach for us would be to instead look at the 'Optimized Logical Plan', and capture everything, acting as a more pure SQL engine. However, the huge disadvantage here is we could actually be less efficient than what Spark would provide by default.

Therefore it makes sense to stick to the idea of rewriting a plan. To do it needs some thinking through, however a really nice part of it is that we can simply rewrite a good chunk of these behaviours into C expressions which would could possibly give us a very elegant solution.