Slim down AuroraSqlPluginTest
There was a lot of repetition and noise as well as a test we no longer need.

In the next change, will look at eliminating some of these tests in favour of some more unified approach. We may be able to, for instance, do some analysis by writing some of our code in the `org.apache.spark.sql` package.

Remove DummyShortCircuitSqlPlugin as it's not really needed.