Compute a word frequency count using C
This C will later be compiled into a VE library.
Effectively we are doing Frovedis-like thing of dual-X86-VE mode.

Frovedis deos quite a bit of stuff however does not support a lot of functionality eg for Strings,
so we need to come up with our own implementations.

Here, we unit test this library using a JNA layer, and later we will transpose it into VE using a custom interface.
This lays an important foundation.