Checkpoint: Arrow word count
We need to find a way, from an arrow representation, to find the offsets, that we can then pass to VEO to process.

We do not want VE to deal with Arrow.

While doing this I found another approach, we could possibly get away with doing this on the Spark side;

That's by using `ArrowColumnVector` to get the address from the underlying `ArrowBuf` which is publicly exposed.