Read from Arrow's `VarCharVector`
This means we can do the following:
Take Spark's InternalRow's, and turn them into a `VarCharVector`, and then pass that through our function to the VE/C kernel.

This way, we avoid a lot of the bureaucracy involved in parsing Arrow stuff. Was complex to get to this point, however this means we can really get to the meat much much quicker!