D cleanup refactor (#69)
This PR doesn't introduce any functionalities but rather solely focuses on refactoring the existing code and unifying some names and so on, since current state made it a little hard to introduce new changes and to reason about what classes are responsible for what. 

I've tried to introduce some naming convention that would clearly state what the classes are actually doing. 

One thing to note is that currently one of the tests `AvgVeSpec` causes `SIGSEGV` but that something that happens for `master` ATM, too so we must have introduced that at some point and I am going to fix taht in a separate PR.