Sort with VE / C++ using Frovedis code (#82)
Frovedis has a lot of VE-optimized code, so we should utilise that. It contains the key logic for Sorts and Joins.

This will allow us to more quickly implement the various functionality especially Hash Joins and such. I will adapt Sorts later as they are a little more complicated to adapt compared to a sort. It is not necessary to implement everything by hand.

We also benefit hugely from C++'s templates that allow a bit of abstraction and generic programming.

We can do this because not all of Frovedis code is tied to MPI. In some cases, we need to modify the code upon import to exclude the MPI bits. Nonetheless, the MPI-independent code is exactly what we need and we should leverage it if we can.