Create a Batch processor, which just duplicates input data.
Our VE based plans will follow more or less this line of execution, because we cannot run 'execute' until we had loaded in all the data.

Nonetheless, we could load the data in incrementally onto the VE, or even asynchronously.

This specific implementation will not necessarily work on large data sets because UnsafeRow might get deallocated.