Abstract Codegen into interfaces we can implement to link with the VE (#102)
`IdentityCodegenBatchPlan` is an example plan that returns back the same rows that were given to it - but after collecting all of them. This utilizes `UnsafeExternalProcessorBase`  and comes with `UnsafeExternalDuplicator` which stores this data. 

**Impact**: we can now develop our own batch processors that call out to Arrow, and then return a result-set.

---

`ArrowSummingCodegenPlan` is a plan that uses `UnsafeExternalProcessorBase` to store the column's data in an Arrow vector, and then sums it up. This requires some benchmarking to see how it compares to eg HashAggregateExec in detail. This one is worth optimizing massively, to call into the VE, to determine if things will work as we expect. If Arrow turns out to be slowing us down then we can do direct memory access.

**Impact**: SUM now works off-heap through direct link with Codegen.

---


`UnsafeExternalProcessorBase` implements a batch load of of data using Codegen, interfacing with an implementation of `UnsafeBatchProcessor`.

**Impact**: as described above. Now we can begin implementing things without strictly relying on Spark internals.

--- 

We should also use `LogicalPlan`, although there are more things to take care of on the inside, on the outside it is actually much simpler to match. We will use `LogicalPlan` as input to generate VE code. This is an important next task to take care of now.

---

This is a culmination of reverse-engineering in the [`w-ve-hash-aggregate`](https://github.com/XpressAI/aurora4spark/commits/w-ve-hash-aggregate) branch, where I explored how to reduce the Codegen complexity of the `BroadcastHashJoinExec`, `HashAggregateExec`, as well as `FilterExec`, `ProjectExec` and other various bits and pieces.

These plans can be combined with Columnar consumption eg of Parquet for much higher performance, I will work on that soon.