Generate JMH benchmarks dynamically
So we can get rich amount of information very quickly, including JFR dumps and such. This is extremely valuable at this point of the project as we really need to focus in on something that is much faster eg than what GPU and Spark can do.

JMH is very difficult to work with because it uses multiple stages of reflections and code generations.
For this reason we generate sources dynamically from the TestingPossibilities file.