Support Projections.
The work here demonstrates that we can return an arbitrary number of rows.
The previous demonstrates we can return an arbitrary number of columns, and output an arbitrary number, as well as perform aggregations.

Next steps feature-wise are to support `GROUP BY` and to support `INNER JOIN`.
Next steps deployment wise are to support integration of the Frovedis libraries into the JAR (rather than referencing them external), so that users can begin to use this. Also we need to integrate the new extension into the AuroraSparkPlugin.

Performance is not going to be amazing, but it works, and we can optimize these different aspects individually at this point.