Do x86 vs VE CSV parsing microbenchmarks, and improve the test quality as well as generic testing. (#124)
The goal here is to isolate Spark away and ensure we are doing the tests fairly.

For the pure-CSV in-memory benchmark, the parsing on VE takes 57 seconds compared to 3s on the CPU for 100M CSV data (3 columns).

For 150MB (to change it around, change it in `Megs(150)` in the files; and then run it with `~ bench -f 1 -wi 0 -prof jfr -i 1 .*NativeCSVParse.*`).

```
[info] Benchmark                                                            Mode  Cnt   Score   Error  Units
[info] DynamicBenchmark.NativeCSVParserBenchmark_CMake_Megs_150_CMake         ss        5.043           s/op
[info] DynamicBenchmark.NativeCSVParserBenchmark_CMake_Megs_150_CMake:·jfr    ss          NaN            ---
[info] DynamicBenchmark.NativeCSVParserBenchmark_VE_Megs_150_VE               ss       85.581           s/op
[info] DynamicBenchmark.NativeCSVParserBenchmark_VE_Megs_150_VE:·jfr          ss          NaN            ---
```

Again this is using my very-streamed CSV reader which is not even very fast at 30MB/s. If we can vectorize we could get a huge performance bump here.

As for the general Spark plans:

The fastest approach we found was VE processing + CMake CSV load. Perhaps GCC is doing some strong optimizations for parsing on x86. Would be interesting to look at what assembly instructions they are using.

```
[info] Benchmark                                                               Mode  Cnt    Score   Error  Units
[info] CSV_CMake_None_NormalCsv_CMake                                            ss       251.478           s/op
[info] CSV_CMake_None_NormalCsv_CMake:·jfr                                       ss           NaN            ---
[info] CSV_PlainSpark_None_NormalCsv_JVM                                         ss       178.254           s/op
[info] CSV_PlainSpark_None_NormalCsv_JVM:·jfr                                    ss           NaN            ---
[info] CSV_Rapids_None_NormalCsv_Rapids                                          ss       193.135           s/op
[info] CSV_Rapids_None_NormalCsv_Rapids:·jfr                                     ss           NaN            ---
[info] CSV_VectorEngine_None_NativeCsvC_VE                                       ss        22.325           s/op
[info] CSV_VectorEngine_None_NativeCsvC_VE:·jfr                                  ss           NaN            ---
[info] CSV_VectorEngine_None_NativeCsvVE_VE                                      ss       361.891           s/op
[info] CSV_VectorEngine_None_NativeCsvVE_VE:·jfr                                 ss           NaN            ---
```

Note benchmarks only run 1ce each, not multiple times, and also there seems to be some variability on the a6 server. Regardless of the variability, the relative performance is quite clear across the different modes.

Since we get a 22s vs 180s for aurora4spark vs plain Spark performance, this is a good result. I am not sure why Rapids is slower than everything else, I would really need somebody to review the performance characteristics here.