Fix a bug in CEvaluationPlan#doExecuteColumnar
It was not reading all the input data - the data was being loaded in set-of-1 partition groups rather than set-of-n partition groups - so the test cases were passing.

This change adds test cases for using parted data across the full stack, as well as extra logging, and fixes the bug - issue was from reusing a pattern in the wrong place.