Implement a deep stack sampler/profiler for JMH
The purpose of this is to identify exactly where time is spent, because there is a constant factor that we might not be aware of.

The output file is a JSON and all the detail how to use this in `SBT.md`

A simple analyzer is included which should be sufficient on the benchmarks. You will need to customize `AnalyzeSamples` (please make PRs with improvements for it eg which files we should ignore) for higher quality output

This should help us find all the funny things that could be happening while our processes are running.