Proof of concept: Java -> UNIX Socket -> VE -> Arrow -> Java
Works on both VE and native C (no Windows)

This shows that we are able to send data incrementally to the VE via a UNIX socket.

The next missing piece is to link up this Socket facility with the Hadoop InputStream reader.

We implemented an outline of the Hadoop InputStream reader, which would produce VE references, that we then pass back to the VE through `CEvaluationPlan`.