Support a single query `SELECT SUBSTR(value, a, b) FROM tbl` - VE String support (#189)
The implementation of `ve_substr` in C is a naive implementation.

As String processing follows a different approach (cannot rely on "evaluate expressions & then VE to vectorize"), we need a new plan and expression evaluation. This will be later replaced with a unified version. Also, because substr can accept both literals and expressions, we need to consider that especially for later on. This is going to demand more funky things to do in terms of code generation: if we have `substr(value, 1)`, it's different from `substr(value, n)` where `n` is another column or an expression.