Use pattern matching properly / reduce more tech debt
We were mixing intents of the data types a little too much. Here we expend in favour of having more precise information about what is possible as an input, as an output and so forth.

Next step will be to create individual unit tests for each of the data types so that we can ensure consistent memory clean-ups and a far cleaner interface. This is not complete however passes the `VectorEngine` tests, so worth to merge it in now. The clean-up is not fully completed but as Wosin is working on NULLs we should get this in first so he has a new stable interface to work with.

A couple of things are removed but we can bring them back if needed.

Strengthened Scala's error checking as well.

![image](https://user-images.githubusercontent.com/52247468/129812804-c1ab08ee-5d64-49c5-a2a6-6524be4455bf.png)