Complete date based filtering
One thing to take away though, it seems that when we do a select l_shipdate, it uses a different plan and pushes down the WHERE_filtering straight to the reader.

We need to think about whether we should actually just allow that to happen, as pre-filtering the data right in Spark-JVM might be faster; we don't really have a VE-based CSV parser here.

Update `build.sbt` to enable live debugging

Add a bit more debug information when transforming plans.