D group by two (#217)
This PR implements Plan and C code for group by sum using multiple columns. It can easily be generalized to any kind of aggregation. The implementation is able to handle nulls correctly too, which is improvement since for previous versions we were just skipping null checks. The results are returned in a different order than spark, because we are sorting the table using both columns and assume null first order, it seems spark does this in a different way, but this probably isn't a problem.