Add support for specifying VE and multiple VEs.
You can specify the VE node to run on by passing VE_NODE_NUMBER as an executorEnv variable.

In the case of YARN with multiple VEs we will assign executors to proper VE address based on the executor ID, wrapping around if more executors are specified than VE cores or if an executor dies and gets the next largest number as the executor ID.

This failure scenario might need more thought.