Prepartition data before aggregation (#299)
This could improve the performance significantly, but requires testing. It avoids a coalesce into a single partition, trading it off for pre-sorting/pre-partitioning data by hashes of the group-by expressions. We also only run 1 function on the VE.

To use:

```
--conf com.nec.spark.preshuffle-partitions=8
```

Otherwise, it will perform as before.

Note: the data ordering is no longer there because of parallelization, and it is not really expected to be sorted - unless we are requesting it to be sorted via `ORDER BY`