`RDD`. `VeColBatch`, `VeColVector`, and an exchange for VE-VE data transfers (#357)
This enables us to simplify the architecture by making it more composable. As part of this, we introduce serialization across boundaries so that different VEs can exchange data via Spark. We also introduce `VeProcess` to wrap AVEO, as well as `executeMulti` which would return multiple distinct datasets (ie multiple columnar batches, in effect).