Be more explicit about what we clear.
We don't want to remove data that has been cached onto the VE.

Our alternate approach could be to cache using Spark's mechanisms, and then deserialize it straight onto the VE using some format VE could understand easily. We should consider this approach for uniformity and also because of super fast transfers that we should be able to attain.