Compile all code in parallel (#394)
For Q1 for example, takes down from 60s to 20s.
For other queries the impact should be even more significant.

The approach is to:
1. Initially pass source code to the plans.
2. At columnar transformation stage, capture all the uncompiled code, compile it, and then `transformUp` to rewrite the VeFunction to a library reference.

This required updating `VeFunction` to support `SourceCode` and `Compiled` modes, as well as creating `PlanCallsVeFunction` which enables us to rewrite the plans.

`ParallelCompilationColumnarRule` is the transformation rule.

Next possible step is to combine all the code into one blob as well, however we should do that in a separate PR.