Rough outline for SubQuery support
This might not work. We might need to possibly rewrite it into a Join instead, because Spark's way of doing things is a little different.

The issue with Subquery is that it may be used in expressions that are beyond conditions - and for that reason we cannot easily join.

So we may have to actually continue onward with this approach.

Work needed here is significant; will try to create a stronger outline but we should focus on minimizing generation of code that is generic as much as possible first.