[group-by] Rewrite the code generation for `GroupingFunction.groupData`
- The original C++ code generation inside `GroupingFunction.groupData` resulted
in compile-able code that came with at least 2 correctnes bugs.  The first was in
the sizing and value assignment for the starts and lens vectors when bucketing
values of nullable_varchar_vector.  The second was that the validity buffer was
not preserved across bucketing.  This resulted garbage results with empty string
keys.  `GroupingFunction.groupData` has been rewritten to generate correct C++
code that addresses these two issues.

- Fix validity buffer preservation across filter operations.

- Temporarily add a query to TPCH benchmarks to continue investigating mismatches
in Query 21 between the JVM and VE modes.