Restore ExchangeOnClusterSpec / distributed runs
Spark uses 2.12.10, and Scala 2.12.15 library changes some internals which get serialized and cannot be deserialized.
This causes issues in distributed runs as if the library uses 2.12.15, and the cluster runners use 2.12.10 then they will receive different things.

Also, make it explicit which Scala version is used and which SPARK home is used in the tests that use distributed Spark