Use another level of indirection to receive arbitrary groups counts
veo_call_sync will copy back only as much data for stack arguments as is set in the allocated veo_args. This puts us into the situation where we can't return an arbitrarily sized array of pointers to results.

To resolve the issue, we introduce an additional layer of indirection. Instead of returning the array of pointers to results, we return just a single pointer to where that array can be found.

As the groups_output result tells us how many results there will be (in addition to their keys), we can use that information to actually request the entire array of output pointers.