Add support for VeColBatch to Tuple conversion
At the interface level we pretend to work on tuples. This in turn means that we need to turn the column vectors back into tuples when we go from VeRDD to regular RDD based execution.

Unfortunately, ClassTag doesn't appear to hold the necessary information for us to know exactly what types are necessary for the tuple conversion. For this reason we move to using TypeTag instead of ClassTag where ever possible.