[NS-39][vectorengine] Implement column vector transfers to and from the VE using the new `VeProcess`
- Add the minimal amount of plumbing to implement column vector transfers to and from the
VE using the new `VeProcess`.  The purpose of this is to enable full functional testing of
the new `VectorEngine` implementation; as such, the transfer code has been duplicated.  Once
the `VectorEngine` has been shown to be functionally correct, work will commence to de-duplicate
this code, integrate the new `VeProcess`, and deprecate the old `VeProcss`

- Add new implementation of `WithVeProcess` that is based on the new `VeProcess` implementation