[NS-56] Remove `RowToColumnar` and `ColumnarToRow` exec nodes
- Prevent Spark from auto-injecting `RowToColumnar` and `ColumnarToRow`
exec nodes into the SQL execution plan.

This is done by first explicitly marking `supportsColumnar` to be `false` across
all `SupportsVeColBatch` plans - we don't actually use `executeColumnar()`which
returns `RDD[ColumnarBatch]`; instead, we use `executeVeColumnar()`, which returns
`RDD[VeColBatch]`.

Next, `VectorEngineToSparkPlan` is updated such that `doExecute()` calls `executeVeColumnar()`
on the child plan (presumably  a `SupportsVeColBatch` plan), and manually converts
`VeColBatch` into a `BytePointerColBatch`, from which we can return an `Iterator[InternalRow]`

We make the returned `Iterator[InternalRow]` lazier by not transposing the columns in JVM, but
by keeping a sequence of iterators, one for each column of the `BytePointerColBatch`