[NS-55] Packed Transfers after Exchange
- Update `VeColBatch.toStream()` and `VeColBatch.fromStream()` to move data to
the VE using packed transfers after VE exchange.  In the new serialization format,
the `UnitColVector` representation of the column vector metadata is serialized,
followed by the transfer buffer bytes.  Upon deserialization, the `UnitColVector`
and buffer are deserialized from the stream and packed into a `UcvTransferDescriptor`
which will be used to transfer the data to the VE.

- Comment out `ColumnarBatchToVeColBatch` and related tests, which are not used anymore