[NS-77] Enhance the build cache implementation used in `CachingNativeCodeCompiler`
- Update `VeFunctionStatus` definitions so that `SourceCode` contains exactly one
`NativeFunction` as opposed to multiple, to make things in line with the "one SparkPlan
one VE function" design used across SparkCyclone

- Enhance the build cache used in `CachingNativeCodeCompiler` with `CompiledCodeInfo` to
contain more information about the compiled function, namely the name of the function.
This is to be used for name replacement in `VeFunction` during the plan transofmration phase in
`CombinedCompilationColumnarRule`, as the semantically-equivalent `NativeFunction` (same
`hashId`) found in the build cache may have been compiled under a different `CFunction2` name.

- Finish the updates to `CombinedCompilationColumnarRule` to support plan transformations for
`VeFunction`s that have `NativeFunction` defined