[NS-77] Redefine the `NativeFunction` abstraction to be more flexible
- Redefine the `NativeFunction` abstraction to contain exactly one primary `CFunction2`
and (optionally) multiple secondary `CFunction2`s, instead of being a 1:1 mapping to
`CFunction2`.  The motivation for this is that while all `VeFunction`s map to exactly
one symbol that will be called by the VE process during the execution of a `SparkPlan`,
the implementation may be complex such that the the function being called will call other
defined functions (VE joins are an example of this). With this change, `NativeFunction`
refers to a function in the semantic sense that is implemented with a primary C function
and possibly multiple secondary helper C functions